+++ lib/mnml/io/in.c
@@ -34,7 +34,7 @@ lisp_function_in(const lisp_t lisp, const atom_t closure)
       fd = dup(0);
       break;
     case T_NUMBER:
-      fd = dup(CHAN->number);
+      fd = dup((int)CHAN->number);
       break;
     case T_PAIR:
       /*
